Refactor project directories and change .scss to .less
