+++ src/components/Apps/AppPreview.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Link } from 'react-router';
-import './AppPreview.scss';
+import './AppPreview.less';
 
 const AppPreview = ({ app }) => {
   return (
