+++ src/components/Apps/Apps.js
@@ -1,8 +1,6 @@
 import React, { Component } from 'react';
-import { Link } from 'react-router';
 import fetch from 'isomorphic-fetch';
-import Icon from '../widgets/Icon';
-import Loading from '../widgets/Loading';
+import Loading from '../../widgets/Loading';
 import AppPreview from './AppPreview';
 
 export default class MyApps extends Component {
@@ -18,12 +16,7 @@ export default class MyApps extends Component {
 
   componentWillMount() {
     this.setState({ isLoading: true });
-
-    fetch(`/api/apps/me`, {
-      headers: new Headers({
-        Authorization: this.props.auth.token,
-      })
-    })
+    fetch('/api/apps')
       .then(res => res.json())
       .then(apps => {
         this.setState({
@@ -32,25 +25,19 @@ export default class MyApps extends Component {
           isLoaded: true,
         });
       });
-  };
+  }
 
   render() {
     const { apps, isLoading, isLoaded } = this.state;
     return (
       <div className="container my-5">
-        <h2>My Apps</h2>
-        <p>These are your apps.</p>
+        <h2>Apps</h2>
         {isLoading && <Loading/>}
         {isLoaded &&
           <div>
             {apps.map((app, key) =>
               <AppPreview app={app} key={key} />
             )}
-            <div className="list-group-item">
-              <Link to="/apps/create">
-                <Icon name="add"/> New App
-              </Link>
-            </div>
           </div>
         }
       </div>
