+++ src/components/Auth/Login.js
@@ -1,8 +1,8 @@
 import React, { Component } from 'react';
 import fetch from 'isomorphic-fetch';
 import { decode } from 'steem/lib/auth/memo';
-import SignForm from '../sign/SignForm';
-import Loading from '../widgets/Loading';
+import SignForm from '../Sign/SignForm';
+import Loading from '../../widgets/Loading';
 
 export default class Login extends Component {
   constructor(props) {
