+++ src/components/Sign/Placeholder/SignPlaceholderComment.js
