+++ src/routes.js
@@ -2,26 +2,26 @@ import React from 'react';
 import { Route, IndexRoute } from 'react-router';
 import Wrapper from './wrapper';
 import Layout from './app';
-import Index from './statics/Index';
-import Dashboard from './statics/Dashboard';
-import App from './apps/App';
-import Apps from './apps/Apps';
-import MyApps from './apps/MyApps';
-import AuthorizedApps from './apps/AuthorizedApps';
-import EditApp from './apps/EditApp';
-import Steemjs from './statics/Docs/Steemjs';
-import OAuth2 from './statics/Docs/OAuth2';
-import Sign from './sign/Sign';
-import RecoverAccount from './components/RecoverAccount';
-import RequestAccountRecovery from './components/RequestAccountRecovery';
-import CreateAccount from './sign/CreateAccount';
-import Login from './oauth2/Login';
-import Authorize from './oauth2/Authorize';
-import RequireLogin from './oauth2/RequireLogin';
-import Permissions from './user/Permissions';
-import SignAuthorize from './sign/Authorize';
-import Revoke from './sign/Revoke';
-import Error404 from './statics/Error404';
+import Index from './components/Index';
+import Dashboard from './components/Dashboard';
+import App from './components/Apps/App';
+import Apps from './components/Apps/Apps';
+import MyApps from './components/Apps/MyApps';
+import AuthorizedApps from './components/Apps/AuthorizedApps';
+import EditApp from './components/Apps/EditApp';
+import Steemjs from './components/Docs/Steemjs';
+import OAuth2 from './components/Docs/OAuth2';
+import Sign from './components/Sign';
+import RecoverAccount from './components/Accounts/RecoverAccount';
+import RequestAccountRecovery from './components/Accounts/RequestAccountRecovery';
+import CreateAccount from './components/Accounts/CreateAccount';
+import Login from './components/Auth/Login';
+import Authorize from './components/Auth/Authorize';
+import RequireLogin from './components/Auth/RequireLogin';
+import Permissions from './components/Accounts/Permissions';
+import SignAuthorize from './components/Accounts/Authorize';
+import Revoke from './components/Accounts/Revoke';
+import Error404 from './components/Error404';
 
 export default (
   <Route path="/" component={Wrapper}>
