+++ src/widgets/Avatar.less
